Theme Colors, Layout and Styling (#146)
* feat: menu bar colors and style, desktop and mobile

* feat: table colors and styles, button styles, new theme colors

* feat: over all styling and layout fixes

* feat: make up fixes

* Cosmetics

Adjusted message, show parent if different from original when cloning

* Add disfunctional LSETH position to blacklist

We should not show the small remaining V1 LSETH position. It has only about 3000 ZCHF left to mint, but the frontend does not show the amount correctly. As V2 has a different way to calculate the limit, I think we should just hide the position for now instead of confronting the user with frontend bugs. Instead of fixing the issue for an obsolete version, we should focus on ensuring the new version runs well.

---------

Co-authored-by: Luzius Meisser <luzius@aktionariat.com>